cargo fix

+++ src/cli/commands/renew.rs
@@ -57,7 +57,7 @@ pub async fn handle_renew(domains: Vec<String>, _force: bool, storage_path: Stri
     let _now = SystemTime::now();
 
     // Get expiration duration
-    let expiry_timestamp = Timestamp::from_millisecond(not_after.timestamp() as i64 * 1000)
+    let expiry_timestamp = Timestamp::from_millisecond(not_after.timestamp() * 1000)
         .map_err(|_| {
             crate::error::AcmeError::Certificate("Failed to parse expiration timestamp".to_string())
         })?;
